+++ SquirrelDefender/appsrc/detect_target.cpp
@@ -1,8 +1,7 @@
 #ifdef ENABLE_CV
-#ifdef _WIN32
 
 /********************************************************************************
- * @file    detect_target_yolo.cpp
+ * @file    detect_target.cpp
  * @author  Cameron Rose
  * @date    6/7/2023
  * @brief   All methods needed to initialize and create a detection network and
@@ -12,7 +11,7 @@
 /********************************************************************************
 * Includes
 ********************************************************************************/
-#include "detect_target_yolo.h"
+#include "detect_target.h"
 
 /********************************************************************************
 * Typedefs
@@ -25,9 +24,6 @@
 /********************************************************************************
 * Object definitions
 ********************************************************************************/
-cv::dnn::Net net;
-std::vector<yolo_net::detection> yolo_detections;
-int yolo_detection_count;
 
 /********************************************************************************
  * Calibration definitions
@@ -38,48 +34,78 @@ int yolo_detection_count;
  ********************************************************************************/
 
 /********************************************************************************
- * Function: YOLO
+ * Function: Detection
  * Description: Class constructor
  ********************************************************************************/
-YOLO::YOLO(void) {};
+Detection::Detection(void) {};
 
 /********************************************************************************
- * Function: ~YOLO
+ * Function: ~Detection
  * Description: Class destructor
  ********************************************************************************/
-YOLO::~YOLO(void) {};
+Detection::~Detection(void) {};
 
 /********************************************************************************
  * Function: initialize_detection_net
  * Description: Delete detection network to free up resources.
  ********************************************************************************/
-bool YOLO::init(void)
+bool Detection::init(void)
 {
-    const std::string class_list_path = "../networks/yolov5m/coco.names";
-    const std::string model = "../networks/yolov5m/yolov5m.onnx";
-    net = yolo_net::create(model, class_list_path, cv::dnn::DNN_BACKEND_CUDA, cv::dnn::DNN_TARGET_CUDA);
+#ifdef JETSON_B01
 
-    return true;
+    return SSD::init();
+
+#elif _WIN32
+
+    return YOLO::init();
+
+#else
+
+#error "Please define a build platform."
+
+#endif
 }
 
 /********************************************************************************
  * Function: detection_loop
  * Description: Process video stream and output detected objects.
  ********************************************************************************/
-void YOLO::loop(void)
+void Detection::loop(void)
 {
-    yolo_net::detect(image, net, yolo_detections);
-    yolo_detection_count = yolo_detections.size();
+#ifdef JETSON_B01
+
+    return SSD::loop();
+
+#elif _WIN32
+
+    return YOLO::loop();
+
+#else
+
+#error "Please define a build platform."
+
+#endif
 }
 
 /********************************************************************************
  * Function: shutdown
  * Description: Shutdown detection network
  ********************************************************************************/
-void YOLO::shutdown(void)
+void Detection::shutdown(void)
 {
+#ifdef JETSON_B01
 
+    return SSD::shutdown();
+
+#elif _WIN32
+
+    return YOLO::shutdown();
+
+#else
+
+#error "Please define a build platform."
+
+#endif
 }
 
-#endif // _WIN32
 #endif // ENABLE_CV
